MNT: adjust inheritance so isinstance(..., cm.ScalarMappable) works
This gives us a path to deprecate ScalarMappable with warnings if we want.
