+++ lib/matplotlib/collections.py
@@ -158,7 +158,7 @@ class Collection(mcolorizer.ColorizingArtist):
             ``Collection.set_{key}(val)`` for each key-value pair in *kwargs*.
         """
 
-        super().__init__(mcolorizer._get_colorizer(cmap, norm, colorizer))
+        super().__init__(self._get_colorizer(cmap, norm, colorizer))
         # list of un-scaled dash patterns
         # this is needed scaling the dash pattern by linewidth
         self._us_linestyles = [(0, None)]
