+++ lib/matplotlib/colorizer.pyi
@@ -78,7 +78,18 @@ class _ColorizerInterface:
     def autoscale_None(self) -> None: ...
 
 
-class ColorizingArtist(artist.Artist, _ColorizerInterface):
+class _ScalarMappable(_ColorizerInterface):
+    def __init__(
+        self,
+        norm: colors.Normalize | None = ...,
+        cmap: str | colors.Colormap | None = ...,
+    ) -> None: ...
+    def set_array(self, A: ArrayLike | None) -> None: ...
+    def get_array(self) -> np.ndarray | None: ...
+    def changed(self) -> None: ...
+
+
+class ColorizingArtist(_ScalarMappable, artist.Artist):
     callbacks: cbook.CallbackRegistry
     def __init__(
         self,
@@ -88,3 +99,7 @@ class ColorizingArtist(artist.Artist, _ColorizerInterface):
     def set_array(self, A: ArrayLike | None) -> None: ...
     def get_array(self) -> np.ndarray | None: ...
     def changed(self) -> None: ...
+    @property
+    def colorizer(self) -> Colorizer: ...
+    @colorizer.setter
+    def colorizer(self, cl: Colorizer) -> None: ...
